refactor(arch): H4 mapper extraction, H6 RQ migration, M6 service split, L3 barrel
H4: extract mapSupabaseOrder into src/services/orderMapper.ts; orderService
    delegates to it and re-exports for backward compat
H6: useOrderValidation now reads from React Query (useOrdersQuery per stage)
    instead of stale Zustand store arrays; store arrays untouched
M6: extract per-row insert loop + mergeAppSettings from mobile-order route
    into src/services/mobileOrderService.ts; route is now parse→auth→call→respond
L3: add src/services/index.ts barrel re-exporting orderService + mobileOrderService

All: npm run type-check clean, biome lint clean.
